Initial commit.
Some basic stuff works at this point.  Given a WSDL, a Binding class
is generated, along with skeleton classes for all the types found in the
WSDL.  Attributes are parsed from complex types, and put in to the generated
type source files.  The Maven mojo can trigger the generation process
from a pom file as part of a build.

I've got enough working now that for a very specific use case, this
generates client side files that work on an Android phone.  I can't commit
the sample project until I can make it not use propietary WSDL's though, so
you'll have to take my word for it.