The beginnings of a response parsing system.
For a limited use case, responses are parsed in to the proper types.  The
only primitive handled right now is String, and when you mix that with
the fact that generation is still incomplete, you can end up with
broken things, but my private sample project has had some success.